more verbose errror handling
