+++ pallets/manta-pay/src/lib.rs
@@ -64,6 +64,7 @@ use crate::types::{
 };
 use alloc::{vec, vec::Vec};
 use core::marker::PhantomData;
+use errors::{ReceiverLedgerError, SenderLedgerError, TransferLedgerError};
 use frame_support::{traits::tokens::ExistenceRequirement, transactional, PalletId};
 use manta_pay::{
     config::{self, utxo::MerkleTreeConfiguration},
@@ -98,6 +99,7 @@ mod mock;
 #[cfg(test)]
 mod test;
 
+pub mod errors;
 pub mod types;
 pub mod weights;
 
@@ -126,17 +128,6 @@ pub mod pallet {
 
     const STORAGE_VERSION: StorageVersion = StorageVersion::new(1);
 
-    /// Suspend execution of MantaPay extrinsics (via tx-pause) due to unexpected internal ledger error.
-    /// Currently the following errors, which are only controlled by our own code, not user input:
-    /// ChecksumError, MerkleTreeCapacityError, VerifyingContextDecodeError
-    pub trait SuspendMantaPay {
-        fn suspend_manta_pay_execution();
-    }
-
-    impl SuspendMantaPay for () {
-        fn suspend_manta_pay_execution() {}
-    }
-
     /// Pallet
     #[pallet::pallet]
     #[pallet::generate_store(pub(super) trait Store)]
@@ -157,9 +148,6 @@ pub mod pallet {
 
         /// Pallet ID
         type PalletId: Get<PalletId>;
-
-        /// Suspends execution of all extrinsics via TxPause
-        type Suspender: SuspendMantaPay;
     }
 
     /// Fungible Ledger Implementation for [`Config`]
@@ -359,11 +347,6 @@ pub mod pallet {
         /// Transfers cannot include amounts equal to zero.
         ZeroTransfer,
 
-        /// Balance Low
-        ///
-        /// Attempted to withdraw from balance which was smaller than the withdrawal amount.
-        BalanceLow,
-
         /// Invalid Serialized Form
         ///
         /// The transfer could not be interpreted because of an issue during deserialization.
@@ -449,166 +432,59 @@ pub mod pallet {
         /// [`InvalidTransfer`](FungibleLedgerError::InvalidTransfer) from [`FungibleLedgerError`]
         PublicUpdateInvalidTransfer,
 
-        /// Internal Ledger Error
-        ///
-        /// This is caused by some internal error in the ledger and should never occur.
-        InternalLedgerError,
+        /// Fungible Ledger Encode Error
+        FungibleLedgerEncodeError,
 
-        /// Encode Error
-        EncodeError,
-    }
+        /// Sender Ledger Fp Encoding failed.
+        SenderLedgerFpEncodeError,
 
-    impl<T> From<InvalidAuthorizationSignature> for Error<T>
-    where
-        T: Config,
-    {
-        #[inline]
-        fn from(_: InvalidAuthorizationSignature) -> Self {
-            Self::InvalidAuthorizationSignature
-        }
-    }
+        /// Sender Ledger [`OutgoingNote`] failed to decode
+        SenderLedgerOutgoingNodeDecodeFailed,
 
-    impl<T> From<InvalidSourceAccount<config::Config, T::AccountId>> for Error<T>
-    where
-        T: Config,
-    {
-        #[inline]
-        fn from(_: InvalidSourceAccount<config::Config, T::AccountId>) -> Self {
-            Self::InvalidSourceAccount
-        }
-    }
+        /// Reciever Ledger Utxo decode failed
+        ReceiverLedgerUtxoDecodeFailed,
 
-    impl<T> From<InvalidSinkAccount<config::Config, T::AccountId>> for Error<T>
-    where
-        T: Config,
-    {
-        #[inline]
-        fn from(_: InvalidSinkAccount<config::Config, T::AccountId>) -> Self {
-            Self::InvalidSinkAccount
-        }
-    }
+        /// Receiver Ledger Wrong Checksum Error
+        ReceiverLedgerChecksumError,
 
-    impl<T> From<SenderPostError<SenderLedgerError>> for Error<T>
-    where
-        T: Config,
-    {
-        #[inline]
-        fn from(err: SenderPostError<SenderLedgerError>) -> Self {
-            match err {
-                SenderPostError::AssetSpent => Self::AssetSpent,
-                SenderPostError::InvalidUtxoAccumulatorOutput => Self::InvalidUtxoAccumulatorOutput,
-                SenderPostError::UnexpectedError(_) => Self::InternalLedgerError,
-            }
-        }
-    }
+        /// Receiver Ledger Merkle Tree Parameters Decoding Error
+        ReceiverLedgerMTParametersDecodeError,
 
-    impl<T> From<ReceiverPostError<ReceiverLedgerError<T>>> for Error<T>
-    where
-        T: Config,
-    {
-        #[inline]
-        fn from(err: ReceiverPostError<ReceiverLedgerError<T>>) -> Self {
-            match err {
-                ReceiverPostError::AssetRegistered => Self::AssetRegistered,
-                ReceiverPostError::UnexpectedError(e) => {
-                    match e {
-                        ReceiverLedgerError::ChecksumError
-                        | ReceiverLedgerError::MerkleTreeCapacityError => {
-                            T::Suspender::suspend_manta_pay_execution();
-                        }
-                        _ => {}
-                    };
-                    Self::InternalLedgerError
-                }
-            }
-        }
-    }
+        /// Receiver Ledger Utxo Accumulator Item Hash Decoding Error
+        ReceiverLedgerUtxoAccumulatorItemHashDecodeError,
 
-    impl<T> From<SenderLedgerError> for TransferLedgerError<T>
-    where
-        T: Config,
-    {
-        #[inline]
-        fn from(err: SenderLedgerError) -> Self {
-            TransferLedgerError::SenderLedgerError(err)
-        }
-    }
+        /// Receiver Ledger Merkle Tree Out of Capacity Error
+        ReceiverLedgerMerkleTreeCapacityError,
 
-    impl<T> From<ReceiverLedgerError<T>> for TransferLedgerError<T>
-    where
-        T: Config,
-    {
-        #[inline]
-        fn from(err: ReceiverLedgerError<T>) -> Self {
-            match err {
-                ReceiverLedgerError::ChecksumError
-                | ReceiverLedgerError::MerkleTreeCapacityError => {
-                    T::Suspender::suspend_manta_pay_execution();
-                }
-                _ => {}
-            };
-            TransferLedgerError::ReceiverLedgerError(err)
-        }
-    }
+        /// Receiver Ledger Field Element Encoding Error
+        ReceiverLedgerFpEncodeError,
 
-    impl<T> From<FungibleLedgerError> for Error<T>
-    where
-        T: Config,
-    {
-        #[inline]
-        fn from(err: FungibleLedgerError) -> Self {
-            match err {
-                FungibleLedgerError::InvalidAssetId(_) => Self::PublicUpdateInvalidAssetId,
-                FungibleLedgerError::BelowMinimum => Self::PublicUpdateBelowMinimum,
-                FungibleLedgerError::CannotCreate => Self::PublicUpdateCannotCreate,
-                FungibleLedgerError::UnknownAsset => Self::PublicUpdateUnknownAsset,
-                FungibleLedgerError::Overflow => Self::PublicUpdateOverflow,
-                FungibleLedgerError::CannotWithdrawMoreThan(_) => Self::PublicUpdateCannotWithdraw,
-                FungibleLedgerError::InvalidMint(_) => Self::PublicUpdateInvalidMint,
-                FungibleLedgerError::InvalidBurn(_) => Self::PublicUpdateInvalidBurn,
-                FungibleLedgerError::InvalidTransfer(_) => Self::PublicUpdateInvalidTransfer,
-                FungibleLedgerError::EncodeError => Self::EncodeError,
-            }
-        }
-    }
+        /// Receiver Ledger Field Element Decoding Error
+        ReceiverLedgerFpDecodeError,
 
-    /// Transfer Post Error
-    pub type TransferPostError<T> = transfer::TransferPostError<
-        config::Config,
-        <T as frame_system::Config>::AccountId,
-        SenderLedgerError,
-        ReceiverLedgerError<T>,
-        TransferLedgerError<T>,
-    >;
+        /// Receiver Ledger Path Decoding Error
+        ReceiverLedgerPathDecodeError,
 
-    impl<T> From<TransferPostError<T>> for Error<T>
-    where
-        T: Config,
-    {
-        #[inline]
-        fn from(err: TransferPostError<T>) -> Self {
-            match err {
-                TransferPostError::<T>::InvalidShape => Self::InvalidShape,
-                TransferPostError::<T>::InvalidAuthorizationSignature(err) => err.into(),
-                TransferPostError::<T>::InvalidSourceAccount(err) => err.into(),
-                TransferPostError::<T>::InvalidSinkAccount(err) => err.into(),
-                TransferPostError::<T>::Sender(err) => err.into(),
-                TransferPostError::<T>::Receiver(err) => err.into(),
-                TransferPostError::<T>::DuplicateMint => Self::DuplicateRegister,
-                TransferPostError::<T>::DuplicateSpend => Self::DuplicateSpend,
-                TransferPostError::<T>::InvalidProof => Self::InvalidProof,
-                TransferPostError::<T>::UnexpectedError(e) => {
-                    match e {
-                        TransferLedgerError::ChecksumError
-                        | TransferLedgerError::VerifyingContextDecodeError(_) => {
-                            T::Suspender::suspend_manta_pay_execution();
-                        }
-                        _ => {}
-                    };
-                    Self::InternalLedgerError
-                }
-            }
-        }
+        /// Receiver Ledger Full Incoming Note Decoding Error
+        ReceiverLedgerFullNoteDecodeError,
+
+        /// Transfer Ledger Wrong Checksum Error
+        TransferLedgerChecksumError,
+
+        /// Transfer Ledger `VerifyingContext` cannont be decoded
+        TransferLedgerVerifyingContextDecodeError,
+
+        /// Transer Ledger Field Element Encoding Error
+        TransferLedgerFpEncodeError,
+
+        /// Transfer Ledger Unknown Asset
+        TransferLedgerUnknownAsset,
+
+        /// Transfer Ledger Proof Error
+        TransferLedgerProofSystemFailed,
+
+        /// Marker Error, tnis error exists for `PhantomData` should never happen
+        Marker,
     }
 
     impl<T> Pallet<T>
@@ -845,41 +721,6 @@ impl<L, R> AsRef<R> for WrapPair<L, R> {
     }
 }
 
-/// Sender Ledger Error
-pub enum SenderLedgerError {
-    /// Field Element Encoding Error
-    FpEncodeError(scale_codec::Error),
-
-    /// Outgoing Note Decoding Error
-    OutgoingNoteDecodeError(scale_codec::Error),
-
-    /// Asset Spent Error
-    ///
-    /// The asset has already been spent.
-    AssetSpent,
-
-    /// Invalid UTXO Accumulator Output Error
-    ///
-    /// The sender was not constructed under the current state of the UTXO accumulator.
-    InvalidUtxoAccumulatorOutput,
-}
-
-impl From<SenderLedgerError> for SenderPostError<SenderLedgerError> {
-    #[inline]
-    fn from(value: SenderLedgerError) -> Self {
-        match value {
-            SenderLedgerError::AssetSpent => Self::AssetSpent,
-            SenderLedgerError::InvalidUtxoAccumulatorOutput => Self::InvalidUtxoAccumulatorOutput,
-            SenderLedgerError::FpEncodeError(err) => {
-                Self::UnexpectedError(SenderLedgerError::FpEncodeError(err))
-            }
-            SenderLedgerError::OutgoingNoteDecodeError(err) => {
-                Self::UnexpectedError(SenderLedgerError::OutgoingNoteDecodeError(err))
-            }
-        }
-    }
-}
-
 impl<T> SenderLedger<config::Parameters> for Ledger<T>
 where
     T: Config,
@@ -953,59 +794,6 @@ where
     }
 }
 
-/// Merkle Tree Parameters Decode Error Type
-pub type MTParametersError = codec::DecodeError<
-    <&'static [u8] as codec::Read>::Error,
-    <config::UtxoAccumulatorModel as codec::Decode>::Error,
->;
-
-/// Utxo Accumulator Item Hash Decode Error Type
-pub type UtxoItemHashError = codec::DecodeError<
-    <&'static [u8] as codec::Read>::Error,
-    <config::utxo::UtxoAccumulatorItemHash as codec::Decode>::Error,
->;
-
-/// Receiver Ledger Error
-pub enum ReceiverLedgerError<T>
-where
-    T: Config,
-{
-    /// Utxo Decoding Error
-    UtxoDecodeError(scale_codec::Error),
-
-    /// Wrong Checksum Error
-    ChecksumError,
-
-    /// Merkle Tree Parameters Decoding Error
-    MTParametersDecodeError(MTParametersError),
-
-    /// Utxo Accumulator Item Hash Decoding Error
-    UtxoAccumulatorItemHashDecodeError(UtxoItemHashError),
-
-    /// Merkle Tree Out of Capacity Error
-    MerkleTreeCapacityError,
-
-    /// Field Element Encoding Error
-    FpEncodeError(scale_codec::Error),
-
-    /// Field Element Encoding Error
-    FpDecodeError(scale_codec::Error),
-
-    /// Path Decoding Error
-    PathDecodeError(scale_codec::Error),
-
-    /// Full Incoming Note Decoding Error
-    FullNoteDecodeError(scale_codec::Error),
-
-    /// Asset Registered Error
-    ///
-    /// The asset has already been registered with the ledger.
-    AssetRegistered,
-
-    /// Type Marker Parameter
-    Marker(PhantomData<T>),
-}
-
 impl<T> From<ReceiverLedgerError<T>> for ReceiverPostError<ReceiverLedgerError<T>>
 where
     T: Config,
@@ -1015,13 +803,6 @@ where
         if let ReceiverLedgerError::AssetRegistered = value {
             Self::AssetRegistered
         } else {
-            match value {
-                ReceiverLedgerError::ChecksumError
-                | ReceiverLedgerError::MerkleTreeCapacityError => {
-                    T::Suspender::suspend_manta_pay_execution();
-                }
-                _ => {}
-            };
             Self::UnexpectedError(value)
         }
     }
@@ -1136,79 +917,6 @@ where
             }
         }
         Ok(())
-    }
-}
-
-/// Verification Context Decode Error Type
-pub type VerifyingContextError = codec::DecodeError<
-    <&'static [u8] as codec::Read>::Error,
-    <config::VerifyingContext as codec::Decode>::Error,
->;
-
-/// Transfer Ledger Error
-pub enum TransferLedgerError<T>
-where
-    T: Config,
-{
-    /// Wrong Checksum Error
-    ChecksumError,
-
-    /// Verifying Context Decoding Error
-    VerifyingContextDecodeError(VerifyingContextError),
-
-    /// Field Element Encoding Error
-    FpEncodeError(scale_codec::Error),
-
-    /// Unknown Asset Error
-    UnknownAsset,
-
-    /// Fungible Ledger Error
-    FungibleLedgerError(FungibleLedgerError),
-
-    /// Sender Ledger Error
-    SenderLedgerError(SenderLedgerError),
-
-    /// Receiver Ledger Error
-    ReceiverLedgerError(ReceiverLedgerError<T>),
-
-    /// Invalid Transfer Shape
-    InvalidTransferShape,
-
-    /// Proof System Error
-    ProofSystemError(ProofSystemError<config::Config>),
-
-    /// Invalid Transfer Proof Error
-    ///
-    /// Validity of the transfer could not be proved by the ledger.
-    InvalidProof,
-
-    /// Type Marker Parameter
-    Marker(PhantomData<T>),
-}
-
-impl<T> From<TransferLedgerError<T>> for TransferPostError<T>
-where
-    T: Config,
-{
-    #[inline]
-    fn from(value: TransferLedgerError<T>) -> Self {
-        match value {
-            TransferLedgerError::InvalidProof => Self::InvalidProof,
-            TransferLedgerError::InvalidTransferShape => Self::InvalidShape,
-            TransferLedgerError::SenderLedgerError(err) => Self::Sender(err.into()),
-            TransferLedgerError::ReceiverLedgerError(err) => Self::Receiver(err.into()),
-            TransferLedgerError::UnknownAsset => Self::InvalidProof,
-            err => {
-                match err {
-                    TransferLedgerError::ChecksumError
-                    | TransferLedgerError::VerifyingContextDecodeError(_) => {
-                        T::Suspender::suspend_manta_pay_execution();
-                    }
-                    _ => {}
-                };
-                Self::UnexpectedError(err)
-            }
-        }
     }
 }
