+++ pallets/manta-pay/src/mock.rs
@@ -20,7 +20,7 @@ use frame_support::{
     traits::{ConstU32, IsInVec},
     PalletId,
 };
-use frame_system::{EnsureRoot, RawOrigin};
+use frame_system::EnsureRoot;
 use manta_primitives::{
     assets::{
         AssetConfig, AssetIdType, AssetLocation, AssetRegistry, AssetRegistryMetadata,
@@ -252,23 +252,6 @@ impl pallet_asset_manager::Config for Test {
     type WeightInfo = ();
 }
 
-pub struct MantaPaySuspensionManager;
-impl crate::SuspendMantaPay for MantaPaySuspensionManager {
-    fn suspend_manta_pay_execution() {
-        let _ = TransactionPause::pause_transactions(
-            RawOrigin::Root.into(),
-            vec![(
-                b"MantaPay".to_vec(),
-                vec![
-                    b"to_private".to_vec(),
-                    b"private_transfer".to_vec(),
-                    b"to_public".to_vec(),
-                ],
-            )],
-        );
-    }
-}
-
 parameter_types! {
     pub const MantaPayPalletId: PalletId = MANTA_PAY_PALLET_ID;
 }
@@ -278,7 +261,6 @@ impl crate::Config for Test {
     type WeightInfo = crate::weights::SubstrateWeight<Self>;
     type PalletId = MantaPayPalletId;
     type AssetConfig = MantaAssetConfig;
-    type Suspender = MantaPaySuspensionManager;
 }
 
 parameter_types! {
