+++ runtime/dolphin/src/assets_config.rs
@@ -184,5 +184,4 @@ impl pallet_manta_pay::Config for Runtime {
     type WeightInfo = weights::pallet_manta_pay::SubstrateWeight<Runtime>;
     type AssetConfig = DolphinAssetConfig;
     type PalletId = MantaPayPalletId;
-    type Suspender = ();
 }
